+++ feature_book_details/src/main/java/com/allsoftdroid/feature/book_details/di/BookDetailsModule.kt
@@ -2,7 +2,7 @@ package com.allsoftdroid.feature.book_details.di
 
 import androidx.lifecycle.SavedStateHandle
 import com.allsoftdroid.audiobook.feature.feature_audiobook_enhance_details.data.network.LibriVoxApi
-import com.allsoftdroid.audiobook.feature.feature_audiobook_enhance_details.data.repository.BestBookDetailsParser
+import com.allsoftdroid.audiobook.feature.feature_audiobook_enhance_details.utils.BestBookDetailsParser
 import com.allsoftdroid.audiobook.feature.feature_audiobook_enhance_details.data.repository.FetchAdditionalBookDetailsRepositoryImpl
 import com.allsoftdroid.audiobook.feature.feature_audiobook_enhance_details.data.repository.NetworkCachingStoreRepositoryImpl
 import com.allsoftdroid.audiobook.feature.feature_audiobook_enhance_details.data.repository.SearchBookDetailsRepositoryImpl
